Update Python VLM example in README (#1178)
Existing example uses an undefined "read_image" function, and using
max_new_tokens in pipe.generate() resulted in an error with latest
nightly.

I updated the example to work out of the box. Makes it a bit longer, but
this section is hidden by default in the README, so it doesn't add to
visual clutter for people just visiting the repo. Also added links to
the relevant samples.

---------

Co-authored-by: Vladimir Zlobin <vladimir.zlobin@intel.com>