Fix excluding stop str from output for some tokenizer (#1676)
For tokenizer of some model like TinyLlama/TinyLlama-1.1B-Chat-v1.0 ,
when detokenize happens token by token, spaces between detokenized word
could be hide. And when we have several stop_str with different length
it could affect result.

For example:
config.stop_strings = {"atoms with a large number of 11 protons",
"ionized"}
prompt = `Why is the Sun yellow?`
answer = `The Sun is yellow because it is made up of hydrogen and helium
atoms, which are composed of atoms with a large number of protons
(hydrogen has 1 proton, helium has 2 protons). These atoms are highly
ionized ...`

in `match_stop_string()` the first stop_str will not match, but then
when the second one will be checked, `decoded_partially_string` will eq
`2protons).Theseatomsarehighlyionized` and the second stop_str will be
stayed in output